Begin the int->intg conversion.
git-svn-id: svn://svn.code.sf.net/p/lush/code@2206 0babfeb2-de4b-461b-a2bf-1ab085750f0d

+++ lush1/trunk/src/storage.c
@@ -24,7 +24,7 @@
  ***********************************************************************/
 
 /***********************************************************************
- * $Id: storage.c,v 1.11 2003-05-27 21:49:50 leonb Exp $
+ * $Id: storage.c,v 1.12 2004-10-22 19:50:36 leonb Exp $
  **********************************************************************/
 
 
@@ -960,7 +960,7 @@ DX(xstorage_realloc_nc)
 #ifdef HAVE_MMAP
 
 void 
-storage_mmap(at *atp, FILE *f, int offset)
+storage_mmap(at *atp, FILE *f, size_t offset)
 {
   struct storage *st;
   size_t len;
